refactoring connectors to separate files
Changes to be committed:
	new file:   src/connectors/BigQueryConnector.py
	new file:   src/connectors/CommonWarehouseConnector.py
	modified:   src/connectors/RedshiftConnector.py
